Secure APIs.

+++ express-ecom-be/app/product/controllers/productController.js
@@ -67,11 +67,22 @@ exports.update = (req, res)=>{
 
 exports.delete = (req, res)=>{
     const productID = req.params.id;
-    repository.delete(productID, (err)=>{
-        if(err){      
-            res.status(HTTPStatus.StatusCodes.INTERNAL_SERVER_ERROR).send(err);     
-        }else{
-            res.status(HTTPStatus.StatusCodes.OK).send("Product is deleted");
+    repository.get(productID, (err, data)=>{
+        if(data){
+            console.log(data);
+            console.log(req.user.sellerid);
+            if(data.sellerID==req.user.sellerid){
+                repository.delete(productID, (err)=>{
+                    if(err){      
+                        res.status(HTTPStatus.StatusCodes.INTERNAL_SERVER_ERROR).send(err);     
+                    }else{
+                        res.status(HTTPStatus.StatusCodes.OK).send("Product is deleted");
+                    }
+                })
+            }else{
+                res.status(HTTPStatus.StatusCodes.UNAUTHORIZED).send("Product doesn't belong to you.")
+            }
         }
     })
+   
 }
